refactor: split up travis/index.js
