+++ lib/utils.js
@@ -22,3 +22,13 @@ exports.getErrorMessage = (error) => {
 
     return null;
 };
+
+
+exports.toArray = (v) => {
+
+    if (v === undefined) {
+        return [];
+    }
+
+    return Array.isArray(v) ? v : [v];
+};
